+++ packages/wavs/src/http/handlers/service/save.rs
@@ -30,16 +30,8 @@ async fn save_service_inner(
     state: HttpState,
     service: wavs_types::Service,
 ) -> HttpResult<SaveServiceResponse> {
-    let service_hash = service.hash()?;
-    if state.load_service(&service_hash).is_ok() {
-        return Err(anyhow::anyhow!(
-            "Service Hash {} has already been set on the http server",
-            service_hash
-        )
-        .into());
-    }
-
-    state.save_service(&service_hash, &service)?;
-
+    // this does NOT save to the dispatcher, it's just for testing purposes, basically simulating IPFS
+    // the url derived from here is typically used to create a ServiceManager instance, e.g. via SetServiceURI
+    let service_hash = state.save_service_by_hash(&service)?;
     Ok(SaveServiceResponse { hash: service_hash })
 }